main_computer/requirements.txt
opencv-python
requests
configparser
zeroconf
pyqt
